adds serialization module, moves types into separate file
